2.0

+++ src/main/java/site/rainbowx/backend/repository/UserRepository.java
@@ -3,10 +3,9 @@ package site.rainbowx.backend.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import site.rainbowx.backend.entity.User;
-
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> { // 修正主键类型为Long
     Optional<User> findByEmail(String email);
 }
